refactor(runtime): change return type of Field::get_field_offset_{xxx} from size_t to uint32_t.

+++ src/runtime/vm/field.cpp
@@ -97,7 +97,7 @@ RtResult<const metadata::RtFieldInfo*> Field::inflate_field(const metadata::RtFi
     RET_OK(&inflatedClass->fields[fieldIndex]);
 }
 
-size_t Field::get_field_offset_includes_object_header_for_all_type(const metadata::RtFieldInfo* field)
+uint32_t Field::get_field_offset_includes_object_header_for_all_type(const metadata::RtFieldInfo* field)
 {
     if (is_instance(field))
     {
@@ -110,7 +110,7 @@ size_t Field::get_field_offset_includes_object_header_for_all_type(const metadat
 }
 
 // Get field offset including object header for reference types
-size_t Field::get_field_offset_includes_object_header_for_reference_type(const metadata::RtFieldInfo* field)
+uint32_t Field::get_field_offset_includes_object_header_for_reference_type(const metadata::RtFieldInfo* field)
 {
     if (is_instance(field) && Class::is_reference_type(static_cast<metadata::RtClass*>(field->parent)))
     {
@@ -123,14 +123,14 @@ size_t Field::get_field_offset_includes_object_header_for_reference_type(const m
 }
 
 // Get field offset including object header for all types
-size_t Field::get_instance_field_offset_includes_object_header_for_all_type(const metadata::RtFieldInfo* field)
+uint32_t Field::get_instance_field_offset_includes_object_header_for_all_type(const metadata::RtFieldInfo* field)
 {
     assert(is_instance(field));
     return field->offset + vm::RT_OBJECT_HEADER_SIZE;
 }
 
 // Get field offset excluding object header
-size_t Field::get_field_offset_excludes_object_header_for_all_type(const metadata::RtFieldInfo* field)
+uint32_t Field::get_field_offset_excludes_object_header_for_all_type(const metadata::RtFieldInfo* field)
 {
     return field->offset;
 }
